visionAI/classificationfashion/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cycler==0.12.1
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.1
pillow==10.4.0
psutil==6.0.0
pyparsing==3.1.4
PyYAML==6.0.2
requests==2.32.3
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
ultralytics==8.3.3
urllib3==2.2.3
Werkzeug==3.0.4
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
importlib_resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
six==1.16.0
zipp==3.19.2
